product-service/requirements.txt
blinker==1.9.0
click==8.3.1
Flask==3.1.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
Werkzeug==3.1.4
motor==3.6.0
pymongo==4.9.1
pydantic==2.9.2
pydantic-settings==2.5.2
python-json-logger==2.0.7
PyJWT==2.8.0
